レスポンスボディの型定義
